refactored before method loader implementation
